fixed helper methods
