+++ spec/rails_helper.rb
@@ -7,6 +7,7 @@ require 'spec_helper'
 require 'rspec/rails'
 require 'capybara/rails'
 require 'capybara/rspec'
+require_relative 'helpers'
 
 ActiveRecord::Migration.maintain_test_schema!
 
@@ -15,20 +16,5 @@ RSpec.configure do |config|
   config.use_transactional_fixtures = true
   config.infer_spec_type_from_file_location!
   config.filter_rails_from_backtrace!
-end
-
-def sign_in_as(user)
-  visit new_user_session_path
-  fill_in 'Login', with: "#{user.username}"
-  fill_in 'Password', with: 'password'
-  click_on 'Log in'
-end
-
-def fill_in_order
-  fill_in 'Cat number', with: 'ABC123'
-  fill_in 'Product name', with: 'Test Product'
-  fill_in 'Quantity', with: '2'
-  fill_in 'Price', with: '15.25'
-  fill_in 'URL', with: 'http://www.google.com'
-  click_on 'Add Item'
+  config.include Helpers
 end
